Move archive handling to common location.
Logic for reading from archive files (e.g. zip files) was part of the
Package Manager module; this commit moves it into the top-level xword
module so it can be reused by other scripts.
